+++ src/getHtml.ts
@@ -154,9 +154,9 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 	let bodyPageHtml = ``
 	 {
 		bodyPageHtml= /*html*/`
-		<div class="page full-h">
+		<div id="full-page" class="page full-h">
 
-			<div class="all-options">
+			<div id="all-options" class="all-options">
 
 				<table>
 					<thead>
@@ -505,7 +505,7 @@ export function createEditorHtml(webview: vscode.Webview, context: vscode.Extens
 
 			</div>
 
-			<div class="table-action-buttons">
+			<div id="table-actions" class="table-action-buttons">
 
 				<div class="separated-btns">