+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+# 0.11.5
+- fixed issue #197 - table resize now works better
+
 # 0.11.4
 - added feature #194 - Better unsaved changes indicator
 - added feature #193 - Copy column header to clipboard